feat: add cards style and some minor changes on layout
