+++ src/App.css
@@ -1,6 +1,42 @@
 body {
   background-color: #282a36;
+  padding-bottom: 100px;
 
+}
+
+.card {
+  border: 1px solid #ccc;
+  padding: 10px;
+  border-radius: 5px;
+  max-width: 95%;
+  margin: 0 auto;
+}
+
+.card-header {
+  border-bottom: #CCC 1px solid;
+  padding: 10px;
+  margin-bottom: 2rem;
+}
+
+.card-header h1 {
+  font-size: 1.5em;
+  text-align: center;
+  color: #FFF;
+}
+
+.card-footer {
+  border-top: #CCC 1px solid;
+  padding: 10px;
+  margin-top: 2rem;
+}
+
+.card-footer a {
+  color: #FFF;
+}
+
+.card-footer i {
+  margin: 0 .5em;
+  color: #FFF;
 }
 
 #root {
