--- src/components/BadPhoneInput/style.module.css
@@ -1,27 +1,0 @@
-.wrapper {
-    border: 1px solid #ccc;
-    width: fit-content;
-    padding: 10px;
-    border-radius: 5px;
-}
-
-.wrapper>h1 {
-    font-size: 1.5em;
-    text-align: center;
-    color: #FFF;
-}
-
-.wrapper>label {
-    display: block;
-    width: 400px;
-    text-align: left;
-    margin: 0 auto 10px;
-    color: #FFF;
-}
-
-.wrapper input {
-    width: 400px;
-    padding: 5px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
