+++ src/components/BrazilianPhone/style.module.css
@@ -1,0 +1,18 @@
+.wrapper>label {
+    display: block;
+    margin-bottom: 10px;
+}
+
+.wrapper input {
+    box-sizing: border-box;
+    width: 100%;
+    padding: 5px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+}
+
+@media screen and (min-width: 820px) {
+    .wrapper {
+        width: 400px;
+    }
+}
